Add files via upload
The aim of this lab is to provide you a refresher on the Pandas library, so that you can pre-process and analyze the datasets before applying data visualization techniques on it. This lab will work as a crash course on pandas. if you are interested in learning more about the pandas library, detailed description and explanation of how to use it and how to clean, munge, and process data stored in a pandas dataframe are provided in our course Data Analysis with Python and Python for Applied Data Science